Add Suspended Status to HTCondorJobManager
If a job is suspended, it will get the "JobStatus == 7" flag. Up to now the status was mapped to `cls.FAILED` and consequently resubmitted. But the job is running again after some time, which leads to multiple submissions of the same job (in the worst case). This change treats the status suspended as `cls.PENDING` instead of `cls.FAILED` in order to avoid resubmission. 
It seems that the suspended status "7" is not an official status from http://pages.cs.wisc.edu/~adesmet/status.html, but definitely present at e.g. RWTH htcondor. Therefore it does not interfere with the "standard" htcondor batch systems.