Fix querying slurm jobs after they're gone
When a slurm job is cancelled out of Law's control and while Law is not
actively polling it, Law still thinks it is running. On the first poll, it will
then normally find out that the job has failed and restart it.

However, the slurm scheduler forgets about old jobs after some time. Trying to
query the status of such a job returns an error. Handle this by returning a
FAILED status for all jobs when squeue fails with "Invalid job id specified".