Add support for optional outputs in TargetCollection
Workflow outputs are wrapped in a TargetCollection. When some of them are
optional and missing, the workflow correctly reports that it's done if all
non-optional outputs are present. However, trying to depend on such a workflow
doesn't work: Luigi checks .exists() before it runs the dependent task, and
dies with "initially missing tasks outputs". To avoid this issue, .exists()
must ignore optional outputs as is done in this patch.

Ran into this when moving the JSON control files for a workflow, which are set
as optional outputs by Law itself. (See #126.)

Didn't touch the file-based collections because I'm not sure how the
`threshold` parameter should be handled (which sounds more likely to be used
than with a mere heterogenous TargetCollection).